Array sequence numbers now 64bit (as per fs.hg / 2744:f02f5bfe6f5e).

+++ src/java/com/acunu/castle/Castle.java
@@ -388,7 +388,7 @@ public final class Castle
 
     public native void castle_merge_thread_destroy(int thread_id) throws CastleException;
 
-    public native int castle_merge_start(int[] array_list, long[] data_ext_list,
+    public native int castle_merge_start(long[] array_list, long[] data_ext_list,
                                          int metadata_ext_type, int data_ext_type,
                                          int bandwidth) throws CastleException;
 
@@ -419,7 +419,7 @@ public final class Castle
         castle_merge_thread_destroy(thread_id);
     }
 
-    public int merge_start(int[] array_list, long[] data_ext_list, int metadata_ext_type,
+    public int merge_start(long[] array_list, long[] data_ext_list, int metadata_ext_type,
                            int med_ext_type, int bandwidth) throws CastleException
     {
         return castle_merge_start(array_list, data_ext_list, metadata_ext_type, med_ext_type,
